do distance maths in c on arduino instead of fiddling with reading in enormous data types

+++ src/asmcnc/production/encoder_connection.py
@@ -138,6 +138,6 @@ class EncoderConnection(object):
     def process_grbl_push(self, message):
 
         if message.startswith('H:'):
-                self.H_side = long(message.split(':')[1])
+                self.H_side = float(message.split(':')[1])
         elif message.startswith('F:'):
-                self.F_side = long(message.split(':')[1])
+                self.F_side = float(message.split(':')[1])
